backend/user_service/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
cryptography
apscheduler
psycopg2-binary

alembic
pytest
httpx
aiosqlite